handle_401: made some code cleanup
