+++ src/main/java/com/ystrazhko/git/connections/RESTConnectorFactory.java
@@ -1,8 +1,10 @@
 package com.ystrazhko.git.connections;
 
+import com.ystrazhko.git.exceptions.HTTPExceptionProvider;
+
 public class RESTConnectorFactory {
 
     public RESTConnector getRESTConnector() {
-        return new RESTConnectorImpl();
+        return new RESTConnectorImpl(HTTPExceptionProvider.getInstance());
     }
 }
